Add comprehensive tests for vectorizer training, cover processing, and album creation workflows
- Implemented tests for the VectorizationService to ensure successful training with tags, error handling, and metadata updates.
- Developed tests for the worker tasks related to vectorization, including tag retrieval and error scenarios.
- Created extensive tests for the improved image cover processing worker, focusing on prioritization, caching, and maintenance tasks.
- Added a workflow demonstration script to illustrate correct and incorrect album creation processes, including error handling and solutions for developers.

+++ backend/library_api/api/schemas/tracks_schema.py
@@ -9,7 +9,7 @@ class TrackBase(BaseModel):
     path: str = Field(..., description="Chemin du fichier")
     track_artist_id: int = Field(..., description="ID de l'artiste principal")
     album_id: Optional[int] = Field(None, description="ID de l'album")
-    duration: Optional[float] = None
+    duration: Optional[int] = None
     track_number: Optional[str] = Field(None, description="Numéro de piste (ex: '01/12')")
     disc_number: Optional[str] = Field(None, description="Numéro de disque (ex: '1/2')")
     year: Optional[str] = Field(None, description="Année de sortie")
